收掉 src/tools 全部 26 行 pending，并搬入 expNibble 的 int 溢出修复
逐个看 LK 的实际 diff 后落账，pending 276→250：

- noise 18 行：LK 的改动纯属风格改写与 `// HanHuaMod` 注释掉的旧代码
  （`int P[]`→`int[] P`、增强 for、javadoc 重排）。tools/data 那 11 个另有一层
  理由：整个 IO 层已被 Cosmic 换成 org/gms/net/packet/*（Netty ByteBuf）
- already-fixed 5 行：LK 的实质改动是把 HeavenMS 原有的 CharsetConstants 多语言
  查表**换成硬编码 GB2312/GBK**（它连 import 一起删了）。BeiDou 保留并改进了查表——
  ByteBufInPacket:56 与 ByteBufOutPacket:79 按 clientLang 取字符集，进出双向一致；
  定长串补齐也在封包层用 writeFixedString 按目标字符集字节数解决。
  LogHelper 的交易日志则已吸收进 Trade.logTrade:588 并接入 i18n
- rejected 1 行：Fishing 把 `(1400.0*rand+1201)*mesoRate` 改成
  `(1400.0*rand+1201*mesoRate)`，倍率只乘常数项，是回退
- deferred 1 行：FilePrinter 新增 LOG_ROYAL / LOG_REDEEM_ROYAL_DAILY 属批次 8
- ported 1 行：MaplePacketCreator 的 (long) 溢出修复搬入 PacketCreator:470。
  getExpNeededForLevel 200 级返回 1,697,021,059，与 int 型 getItemExp 相乘时
  在 int 内溢出后才拓宽成 long，itemExp ≥ 2 即触发

mvn -pl gms-server compile 通过；gen-port-manifest.sh 复验回写完整保留。

+++ gms-server/src/main/java/org/gms/util/PacketCreator.java
@@ -467,7 +467,9 @@ public class PacketCreator {
         } else {
             int itemLevel = equip.getItemLevel();
 
-            long expNibble = (ExpTable.getExpNeededForLevel(ii.getEquipLevelReq(item.getItemId())) * equip.getItemExp());
+            // 两个操作数都是 int，不先转 long 的话乘法在 int 里溢出后才拓宽。
+            // getExpNeededForLevel 200 级返回 1,697,021,059，itemExp 只要 ≥ 2 就爆。
+            long expNibble = ((long) ExpTable.getExpNeededForLevel(ii.getEquipLevelReq(item.getItemId())) * equip.getItemExp());
             expNibble /= ExpTable.getEquipExpNeededForLevel(itemLevel);
 
             p.writeByte(0);
